+++ app/src/main/java/woowacourse/shopping/data/cart/CartProductDao.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data
+package woowacourse.shopping.data.cart
 
 import androidx.room.Dao
 import androidx.room.Insert
