+++ app/src/main/java/woowacourse/shopping/data/mapper/ProductMapper.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.data.mapper
 
-import woowacourse.shopping.data.CartProductEntity
+import woowacourse.shopping.data.cart.CartProductEntity
 import woowacourse.shopping.model.Product
 
 fun Product.toEntity(): CartProductEntity {
