+++ app/src/main/java/woowacourse/shopping/di/module/RepositoryModule.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.di.module
 
-import woowacourse.shopping.data.DefaultCartRepository
 import woowacourse.shopping.data.DefaultProductRepository
+import woowacourse.shopping.data.cart.DefaultCartRepository
 import woowacourse.shopping.repository.CartRepository
 import woowacourse.shopping.repository.ProductRepository
